[ADD] Uncentered gradient covariance matrix (aka 'empirical Fisher') (#17)
* [DOC] Add mathematical definitions of GGN and Hessian

* [FIX] Set dtype of random vector in deterministic check

* [ADD] Linear operator for empirical Fisher

* [DEL] Remove autograd double-check in functorch EF implementation

* [DOC] Add empirical Fisher to visual tour

* [DEL] Remove TODO after failed attempt with `is_grad_batched`

* [FIX] Remove unused import

* [DOC] Add missing docstrings

+++ curvlinops/_base.py
@@ -154,7 +154,7 @@ class _LinearOperator(LinearOperator):
             self.print_nonclose(grad1, grad2, rtol, atol)
             raise RuntimeError("Check for deterministic gradient failed.")
 
-        v = rand(self.shape[0])
+        v = rand(self.shape[0]).astype(self.dtype)
         mat_v1 = self @ v
         mat_v2 = self @ v
